# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cauteum Gateway REST API
  version: 0.1.0
  summary: Client-facing contract for the gateway control-plane REST API.
  description: >-
    This specification documents the client-facing REST control-plane surface:
    health, gateway metadata, sandbox registry, logs, policy proposals,
    providers, workspaces, services, templates, settings, inference, and local
    or OIDC authentication configuration. Operational probes, diagnostics,
    streaming relay transports, compatibility aliases, and RPC are excluded.
    Sandbox registry writes do not provision or stop a compute runtime; runtime
    lifecycle operations are defined by the OpenShell RPC contract.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://127.0.0.1:7443
    description: Local development gateway. Configure HTTPS for remote use.
...
security:
  - bearerAuth: []
paths:
...
  /v1/sandboxes/{name}/ssh-session:
    parameters:
      - $ref: '#/components/parameters/SandboxName'
    post:
      operationId: createSshSession
      tags: [Sandboxes]
      summary: Create an SSH session for a sandbox
      description: Requires a connected supervisor. The session token is returned once and must be handled as a secret.
      responses:
        '200':
          description: SSH connection details and one-time session token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSHSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          description: Sandbox relay is not connected.
          content:
            text/plain:
              schema:
                type: string
...
  /v1/ssh-sessions:
    get:
      operationId: listSSHSessions
      tags: [Sandboxes]
      summary: List SSH sessions
      parameters:
        - name: sandbox
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Active and revoked session records.
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SSHSessionInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/ssh-sessions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: revokeSSHSession
      tags: [Sandboxes]
      summary: Revoke an SSH session
      responses:
        '204':
          description: Session revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or local gateway token
  parameters:
    SandboxName:
      name: name
      in: path
      required: true
      description: Sandbox registry name.
      schema:
        type: string
  schemas:
...
    SSHSessionInfo:
      type: object
      required: [id, sandbox, created_at_ms]
      properties:
        id:
          type: string
        sandbox:
          type: string
        subject:
          type: string
        created_at_ms:
          type: integer
          format: int64
        expires_at_ms:
          type: integer
          format: int64
        revoked:
          type: boolean
...
    SSHSession:
      type: object
      required: [session_id, sandbox_id, token, gateway_scheme, gateway_host, gateway_port, expires_at_ms]
      properties:
        session_id:
          type: string
        sandbox_id:
          type: string
        token:
          type: string
          format: password
          description: Short-lived bearer credential. Store securely and never log it.
        gateway_scheme:
          type: string
          enum: [http, https]
        gateway_host:
          type: string
        gateway_port:
          type: integer
        expires_at_ms:
          type: integer
          format: int64
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token. Current response body is text/plain.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        text/plain:
          schema:
            type: string
    Forbidden:
      description: Caller is not authorized. Current response body is text/plain.
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: Resource not found. Current response body is text/plain.
      content:
        text/plain:
          schema:
            type: string
...